fix: change visibility
